Allow datum dependant durations to be converted to milliseconds by using the current instant as the datum
Also work round java.sql.Date sillyness


git-svn-id: http://svn.codehaus.org/groovy/trunk@3847 a5544e8c-8a19-0410-ba12-f9af4593a198

+++ groovy/modules/gdata/src/main/groovy/time/Duration.java
@@ -81,7 +81,7 @@ public class Duration extends BaseDuration {
     }
     
     public long toMilliseconds() {
-        return ((((((this.getDays() * 24 ) + this.getHours()) * 60 + this.getMinutes()) * 60) + this.getSeconds()) * 1000) + this.getMillis();
+        return ((((((long)(this.getDays() * 24 ) + this.getHours()) * 60 + this.getMinutes()) * 60) + this.getSeconds()) * 1000) + this.getMillis();
     }
     
     public Date getAgo() {
